# neo-backend/internal/handlers/json.go
package handlers

import (
	"encoding/json"
	"net/http"
)

const maxBodyBytes = 1 << 20 // 1MB

// ErrorResponse is the JSON body returned for every non-2xx response.
type ErrorResponse struct {
	Error string `json:"error" example:"invalid request body"`
}

func writeJSON(w http.ResponseWriter, status int, payload any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	json.NewEncoder(w).Encode(payload)
}

func decodeJSON(w http.ResponseWriter, r *http.Request, dst any) error {
	r.Body = http.MaxBytesReader(w, r.Body, maxBodyBytes)
	dec := json.NewDecoder(r.Body)
	dec.DisallowUnknownFields()
	return dec.Decode(dst)
}

func respondError(w http.ResponseWriter, status int, msg string) {
	writeJSON(w, status, ErrorResponse{Error: msg})
}